EnemyIssues and some adjustments
So, now the enemy is attacking each player individually, also " changed  " the tutorial.tscn scene layout. Furthermore, now if the player dies it goes back to the main menu, I tried to add a panel, but the position of the panel does not follow the player, gonna try to fix it later.